panel: Make quick settings configurable in session mode
Right now the gnome-initial-setup mode has quick settings that aren't
that useful (like "Power Mode" and "Dark Style"). It also has some
that are useful (like "Wi-Fi").

Ths commit allows gnome-initial-setup to decide which settings to show
by adjusting its session mode configuration.

This will also give us the flexibility later to evaluate and change
which quick settings are shown in the lock screen, as well.

The changes are 4-fold:

1. Updating the session mode configuration to include a new
quickSettings property which contains the list of quick settings to
load for the given mode. For backward compatibility, by default, this
property is null which means "load the default set of quick settings".

2. A new QuickSettingsManager class that's akin to ComponentManager but
for quick settings instead of components. It handles importing the
status modules that provide system indicators and quick setting menu
items.

3. Changing the QuickSettings panel menu button to use the
QuickSettingsManager instead of importing the status modules itself.

4. Changing the modules that provide quick settings that are supposed to
span multiple columns in the grid to announce that requirement by
setting a new "expand" propery on the modules indicator quick settings
items array.