shell/app: Adapt to focus changes in mutter
meta_window_raise_and_make_recent() now only works as advertised
for windows on the active workspace.

That results in the following unexpected behavior:
 1. open two windows of an app (say: terminal)
 2. open another app, maximize the window for best effect
 3. switch to another workspace
 4. super+tab to the first app: both windows are raised
 5. close the window; instead of moving focus to the
    2nd app window, both windows disappear because focus
    is moved to the app from 2

This was fixed with new mutter API on the main branch, but
for the stable branch it is easier to work around the issue.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/2959>