docs: Elaborate on supported versions
The link to the GNOME schedule was added because of some confusion
of which versions are supported upstream. Briefly outline the
policy of "last two stable + current main", so that the information
can also be deduced from tags etc. without scrolling through the
schedule calendar.

While at it, update the link to the schedule to point to the
future release.gnome.org page.

Death to the wiki!

Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/3122>