docs: Document C coding style
We have a fair bit of C code, yet don't document the expected
coding style. This is not much of an issue for regular GNOME
contributors as we are following the conventions of the wider
project (glib, gtk, clutter, ...), but it's still better to
spell it out.

Do that by adapting the corresponding documentation from mutter.

Even if much of our C code doesn't match all the new guidelines,
it's good to consolidate the coding style between both projects.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/3122>