gdm: Add new AuthMenuButton control
The latest login screen designs show a new "Login Options" menu in the
corner for session selection and login methods.

As a first step toward acheiving that goal, this commit adds a new
AuthMenuButton class. It's a fairly generic control derived from
the code used to show the sessions menu button. One key difference
is it allows a multi-valued, partial key for its entries. This provides
flexibility that we'll need to leverage later.

Also add AuthMenuButtonIndicator which derives from AuthMenuButton to
display icons to inform about non selectable background authentication
methods in use, it's non-interactive.

Nothing uses these new classes yet. A subsequent commit will change the
sessions menu button code over to use it, and a commit after that will
use it for Login Options.